+++ src/components/MainBlock/Calendar/Сalendar.css
@@ -29,6 +29,7 @@
   justify-content: center;
   border: 1px solid #000000;
   color: #000000;
+  transition: background-color 0.3s ease; /* Переходы для цвета */
 }
 
 .empty {
@@ -47,3 +48,8 @@
 .future {
   background-color: bisque;
 }
+
+.day:hover {
+  background-color: rgba(255, 255, 255, 0.5); /* Эффект наведения */
+}
+
