+++ src/components/NotificationBlock/elements/NotiElements/NotiElements.tsx
@@ -6,7 +6,6 @@ const NotiElements = () => {
             <h1 className='message'> У Вас новая бронь!</h1>
             <h2 className='order'> Столик забронирован на 17:00</h2>
         </div>
-      // TODO: Форматирование
     ); 
 };
